Split dataloader into a helper implementation for most of the GUTS of the code
